Move video/video_comment tests to the new folder structure
