+++ tests/Context/Video/Module/VideoComment/Application/Publish/PublishVideoCommentCommandStub.php
@@ -2,12 +2,15 @@
 
 declare(strict_types=1);
 
-namespace CodelyTv\Test\Context\Video\Module\Video\Application\Create;
+namespace CodelyTv\Test\Context\Video\Module\VideoComment\Application\Publish;
 
 use CodelyTv\Context\Video\Module\Video\Domain\VideoId;
 use CodelyTv\Context\Video\Module\VideoComment\Contract\PublishVideoCommentCommand;
 use CodelyTv\Context\Video\Module\VideoComment\Domain\VideoCommentContent;
 use CodelyTv\Context\Video\Module\VideoComment\Domain\VideoCommentId;
+use CodelyTv\Test\Context\Video\Module\Video\Domain\VideoIdStub;
+use CodelyTv\Test\Context\Video\Module\VideoComment\Domain\VideoCommentContentStub;
+use CodelyTv\Test\Context\Video\Module\VideoComment\Domain\VideoCommentIdStub;
 use CodelyTv\Test\Stub\UuidStub;
 use CodelyTv\Types\ValueObject\Uuid;
