+++ tests/Context/Video/Module/VideoComment/Domain/VideoCommentContentStub.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CodelyTv\Test\Context\Video\Module\Video\Domain;
+namespace CodelyTv\Test\Context\Video\Module\VideoComment\Domain;
 
 use CodelyTv\Context\Video\Module\VideoComment\Domain\VideoCommentContent;
 use CodelyTv\Test\Stub\TextStub;
