+++ tests/Context/Video/Module/VideoComment/Domain/VideoCommentIdStub.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CodelyTv\Test\Context\Video\Module\Video\Domain;
+namespace CodelyTv\Test\Context\Video\Module\VideoComment\Domain;
 
 use CodelyTv\Context\Video\Module\VideoComment\Domain\VideoCommentId;
 use CodelyTv\Test\Stub\UuidStub;
