+++ tests/Context/Video/Module/VideoComment/Domain/VideoCommentStub.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CodelyTv\Test\Context\Video\Module\Video\Domain;
+namespace CodelyTv\Test\Context\Video\Module\VideoComment\Domain;
 
 use CodelyTv\Context\Video\Module\Video\Domain\VideoId;
 use CodelyTv\Context\Video\Module\VideoComment\Domain\VideoComment;
